requirements.txt
image == 1.5.33
pillow == 11.1.0
pip == 25.0.1
